requirements.txt
stable-baselines3[extra]
gymnasium
pyyaml
pandas
shapely
traci # Note: Usually comes with SUMO, but can be installed via pip if needed